make content_overrides in k_activation_key truly idempotent
before we considered all results of the product content call as "current
overrides", even when they were not actually set to be overriden. this
resulted in too many "reset to default" calls (for overrides that
already were at "default") plus it made the module return changed=True
in too many cases.

let's only look at the entries that have a not None
enabled_content_override, which means they are actually overriden.

Fixes: #186