Adding locked state to ptable (#135)
* Adding locked state to ptable

Requires:
- https://github.com/theforeman/foreman/pull/5741
- https://github.com/SatelliteQE/nailgun/pull/503

* Remove default value for locked state

Remove the default value for the locked state otherwise the locked state
will be set to "false" in case the "locked" option is not given. Set the
locked state to true/false if you want to change the locked state. In
case of a new template, the locked state is set to false automatically
as this is the default on the foreman side.

* Update version dependency of nailgun

* Adding tests for locked state