Create foreman_user module
Create a new module for managing users. In order to change password we
need to force the `password` field to be sent during POST but that
field is not included when doing GET, so we put it in `check_missing`.

I noticed, though, that the value for `check_missing` was a variable
from a separate loop which seems wrong. Instead I change it to get the
value from the incoming desired dict. This makes the module behave as
we want.

To ensure that the new user works we do a minor API request as the
newly created user.