name of the host should be an fqdn
extracting the domain part from the hostgroup is non deterministic and requires the hostgroup to be specified and to have a domain defined even if module is run with ```state: absent```. Perhaps it's better to add another parameter e.g. ```shortname``` that would require hostgroup and make them require_one_of.