Port foreman_host to apypie and add build parameter (#429)
* ported from nailgun to apypie

* added build option

this will work if the hostgroup (required for now) has all the necessary context to provision the host

* name of the host should be an fqdn

extracting the domain part from the hostgroup is non deterministic and requires the hostgroup to be specified and to have a domain defined even if module is run with ```state: absent```. Perhaps it's better to add another parameter e.g. ```shortname``` that would require hostgroup and make them require_one_of.

* Improved documentation and examples, facelift.

* homogenize bools in documentation string

* add build to tests for host and fix hostname

Fixes #402 
Fixes #426