don't scope host update requests by Organization/Location
Foreman/Apipie allows the same short parameter name to be used multiple
times in an API definition and we fill them identically, assuming they
mean the same.
However, this is not true. For example, when updating a Host, the "outer"
organization_id means to "scope" the request by that ID, while the
"inner" says that the Host is to be created in said Organization.
This makes it impossible to move a host between Organizations, as the
request would be scoped by the new Org, and then Foreman won't find the
host in that Org to update it.

Let's just not scope those update requests.