Allow disabling callback plugin even if it gets loaded
Foreman_ansible decides whether the callback should be used to prevent facts
from being uploaded in case ansible is used "just" as a remote execution
provider. It was implemented by setting ANSIBLE_CALLBACK_WHITELIST environment
variable, but that stomps over any user-configured callback plugins[2]. When we
moved to ansible-runner the whitelist env trick was dropped, but also led to
facts being uploaded even if we don't want them[1].

While we could read the configured callback plugins from ansible's configuration
and change it on the fly, it would be a bit fragile. The configuration can be in
different places and noone guarantees that the location or its contents won't
change in the foreseeable future.

This change allows us to selectively disable the callback plugin if we need to,
without any undesired side-effects.

[1] - https://projects.theforeman.org/issues/32020
[2] - https://projects.theforeman.org/issues/30373