Add a role to perform basic setup to register and patch RHEL clients (#1135)
* Add a role to perform basic setup to register and patch RHEL clients

* Don't match against manifest in content_rhel_role fixture

* Record test fixtures

* Add changelog fragment for content_rhel role

* Skip check_mode tests for content_rhel_role

content_rhel_role uses async tasks, which are unsupported by check_mode

* Trim extraneous requests from test fixtures

The fixtures contain a lot of requests checking on the status of a
Foreman task which is in state running/pending. We can minimize the
diff and speed up execution by eliminating redundant requests/responses
after the first one.