find operatingsystems by title or full (name,major,minor) tuple
Operatingsystems in Foreman have a rather weird uniqueness construct:
they have a `description` attribute, that is unique, but can be blank,
and they are unique by the tuple (`name`, `major`, `minor`) but here
again `minor` can be blank too.
They also have a generated `title` attribute, that is either the value
of the description attribute (if that is not blank) or a concatenation
of name/major/minor, shortened to 255 characters.

In Rails, this looks like that (roughly quoting
https://github.com/theforeman/foreman/blob/develop/app/models/operatingsystem.rb):

    validates :name, :presence => true, :no_whitespace => true,
              :uniqueness => { :scope => [:major, :minor], :message => N_("Operating system version already exists")}
    validates :description, :uniqueness => true, :allow_blank => true
    before_validation :downcase_release_name, :set_title, :stringify_major_and_minor
    validates :title, :uniqueness => true, :presence => true

    def release
      "#{major}#{('.' + minor.to_s) if minor.present?}"
    end

    def fullname
      "#{name} #{release}"
    end

    # The OS is usually represented as the concatenation of the OS and the revision
    def to_label
      return description if description.present?
      fullname
    end

    def set_title
      self.title = to_label.to_s[0..254]
    end

Overall, this construct makes finding a specific OS rather "interesting":
If you have two OSes, "EL 8" and "EL 8.6" (so both have name:EL, major:8
and one has minor:6), to find "EL 8" you can't search for
"name=EL,major=8" as that would return *both*. You could search for
"name=EL,major=8,minor=''" or for "title='EL 8'" and get the unique OS
you wanted.
Yet, when we ask the user to provide a description, they can use more
than 255 characters that is allowed for the title -- so we need to look
up in both fields, title AND description as the former is always there
(generated) while the later will have the exact string *if* set.