redhat_manifest - set default ``quantity`` to 1
The code decides whether subs need to be attached or not by the presence
of the `pool_id` parameter. However, to be able to actually attach a
pool, one also needs to pass the quantity.

We could document that and set a `required_together` property, or do the
sensible thing and just default the `quantity` to 1, which is a sane
default in the days of SCA where we need to attach a sub to make Katello
aware of the fact that we have access to said content, but otherwise
don't need the *correct amount* of subs attached for providing them to
clients.