fetch all products for an AK when managing content overrides
we need to explicitly pass `per_page` to the call, as otherwise the API
will return just the default `per_page` numer (20 or so), thus making
the module not idemtpotent when there are more repositories to be
managed.

sadly needs an apidoc patch, because the param is not exposed yet.

Fixes: https://bugzilla.redhat.com/show_bug.cgi?id=2134605