ensure strings are strings when converting parameters for the API
We define the `value` attribute as `raw` as it can have different data
types, which means Ansible will helpfully convert a `"['something']"`
string to an actual array with the entry `'something'`.
However, if the `parameter_type` in Foreman is `string`, we really
should send a string, not an array Ansible converted for us.